3380: merge from lp:~wbrowa/openpetraorg/dev_20161201 https://tracker.openpetra.org/view.php?id=5725 Tags: BETA1_0.6.22

+++ csharp/ICT/Petra/Server/lib/MPartner/web/Contacts.cs
@@ -406,7 +406,11 @@ namespace Ict.Petra.Server.MPartner.Partner.WebConnectors
                             // now we also need to delete the contact attributes (linked with this contact log)
                             PPartnerContactAttributeRow template = new PPartnerContactAttributeTable().NewRowTyped(false);
                             template.ContactId = ContactLogId;
-                            PPartnerContactAttributeAccess.DeleteUsingTemplate(template, null, Transaction);
+
+                            if (PPartnerContactAttributeAccess.CountUsingTemplate(template, null, Transaction) > 0)
+                            {
+                                PPartnerContactAttributeAccess.DeleteUsingTemplate(template, null, Transaction);
+                            }
 
                             // and the contact log itself needs to be deleted (if no other partner refers to it)
                             PContactLogAccess.DeleteByPrimaryKey(ContactLogId, Transaction);
